refactor(database): score and simplify provider boundaries
